requirements.txt
numpy
pandas
matplotlib
seaborn
tqdm
torch
torchvision
torchaudio
lightning
torchsummary
pyyaml
python-dotenv
pillow
opencv-python
albumentations
insightface
onnxruntime
# onnxruntime-gpu
rich
ipykernel